Add /etc/os-release
Since /etc/os-release is a symlink to /usr/lib/os-release, let's copy
that one as well.

This is to make the Docker image work well in GitHub Action.

The most common step in the GitHub Actions workflows is the one that
checks out the repository: actions/checkout@v3.
The first thing it does is grepping the /etc/os-release file, and if it
does not exist it fails.

Signed-off-by: Vicente Olivert Riera <vincent.olivert.riera@gmail.com>